Issue #94: Move CC's into api and fix factory
